don't explode if network scripts aren't there

+++ board/cm3/post-build.sh
@@ -7,7 +7,18 @@ BOARD_DIR="$(dirname $0)"
 cp -f ${BOARD_DIR}/config.txt $BINARIES_DIR/rpi-firmware/
 
 # Remove autoenabling of networking
-#rm ${TARGET_DIR}/etc/init.d/S40network
-rm ${TARGET_DIR}/etc/init.d/S05avahi-setup.sh
-rm ${TARGET_DIR}/etc/init.d/S41dhcpcd
-rm ${TARGET_DIR}/etc/init.d/S80dnsmasq
+if [ -e ${TARGET_DIR}/etc/init.d/S40network ]; then
+   rm ${TARGET_DIR}/etc/init.d/S40network
+fi
+
+if [ -e ${TARGET_DIR}/etc/init.d/S05avahi-setup.sh ]; then
+    rm ${TARGET_DIR}/etc/init.d/S05avahi-setup.sh
+fi
+
+if [ -e ${TARGET_DIR}/etc/init.d/S41dhcpcd ]; then
+    rm ${TARGET_DIR}/etc/init.d/S41dhcpcd
+fi
+
+if [ -e ${TARGET_DIR}/etc/init.d/S80dnsmasq ]; then
+    rm ${TARGET_DIR}/etc/init.d/S80dnsmasq
+fi
